requirements.txt
google-cloud-bigquery-storage>=2.22,<3
google-auth>=2.25,<3
protobuf>=4.25,<7
